pyproject.toml
[project]
name = "vlm-cot-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "google-genai>=1.56.0",
    "openai>=2.9.0",
    "pillow>=12.0.0",
    "vllm>=0.12.0",
]
